First version of ensemble training script
Currently, this *only* compares the baseline trained on *all* samples
with a classifier trained only on species-specific examples. Still have
to implement a way to perform proper sampling.

(this will require additional changes to `maldi-learn` because we need
to return the stratification vector in order to perform proper sampling
that is aware of class labels)