Simple job resubmission script
Currently, this is only meant for the ensemble jobs, but it can be
re-purposed to all other kinds of jobs. The idea is to *resubmit*,
thus hoping for a second chance.

This makes sense because the ensemble jobs do *two* configurations
and one of them might run out of time. It will just be skipped the
next round.