Added support for automated model subdirectory generation
This will automatically sort the results for each experiment into the
required folder structure. The upside of this is that many scenarios,
or runs, will be skipped automatically now if the data is present. It
is possible to override this behaviour using `--force` in most of the
run scripts, though.